:recycle: split ocr module to match global SDK syntax
