--- mindee/parsing/common/mvision_v1.py
@@ -1,2 +1,0 @@
-class MVisionV1:
-    """Internal OCR Service class."""
