+++ mindee/parsing/common/ocr/__init__.py
@@ -1,0 +1,1 @@
+from mindee.parsing.common.ocr.ocr import Ocr
